remove crawl return value
crawl used to return a map that mapped page urls to the words found on
that page, and this map was used to build the index. crawl was then
refactored to build the index itself. so the map is no longer needed.
I'll also be working on concurrency soon, which I think will make it
much harder to build this map.