perf: extract returns a map from word to count
previously, it gave a slice of words that were in the page. we then
incremented the count of each word when building the index. this meant
that if a word occurred twice, we'd have to increment it twice. by
returning a map with the count, we only have to set it once. this is a
pretty big performance boost for sql especially, since we're reducing
the file writes. this reduces the top10 crawl time from 207.49 seconds
(1.61 per second) to 101.10 seconds (3.31 per second)